LookUpTable Generator for PWM set as function of Desired Inspiration Ramping Speed.
LookUpTable Generator for PWM set as function of Desired Inspiration Ramping Speed. It can be used to tune PWM response as a function of desired arm speed in inspiration. Open Loop - first aproach, it will need tuning. But it is already working.
It could be used as a generic LookUpTable generator for other purposes.
As it was built using arduino IDE, there is no need for another compiler, so Arduino Pro Mini is the LookUpTable generator robot.