Added External WatchDog Timer - Schematics, boards and Gerbers
In order to achieve a reasonable level of reliability, a planned external watchdog was developed. Arduino bootloader is not compatible with watchdogs so this part needs to be removed when upgrading eAR firmware. eAR will be able to recognize when this part is connected due to a MISO MOSI loopback for testing purposes.