Added external watchdog support
As planned, an external watchdog was added to eAR. Arduino is not reliable enough so this is a MANDATORY add-on. It is focused on restarting when arduino firmware loses control. Signals are generated to charge-pump a capacitor. This capacitor is part of an oscillator and when charge is not enough the oscillation resets eAR system. It is necessary remove this add-on when reprogramming arduino. It is socket-mounted on last expansion board available.