added Pressure Sensor Board
The pressure sensor could be integrated in ECU Board but we preferred to have a three-wire cable instead of a long tube so we designed a new board connected to ECU by a 4-way Molex MiniFit Jr connector. The Px Sensor is the 5V NXP MPX5010DP and we use only the upper (pressure) port.